Remove unnecessary cast from long to long.
git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/io/trunk@491359 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/commons/io/EndianUtils.java
@@ -224,7 +224,7 @@ public class EndianUtils {
             ( ( data[ offset + 5 ] & 0xff ) << 8 ) +
             ( ( data[ offset + 6 ] & 0xff ) << 16 ) +
             ( ( data[ offset + 7 ] & 0xff ) << 24 );
-        return ((long) high << 32) + (0xffffffffL & low); 
+        return (high << 32) + (0xffffffffL & low); 
     }
 
     /**
